pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.2.0"
description = ""
authors = ["VaddyVG <vads1999@inbox.ru>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"

[tool.pytest.ini_options]
addopts = "--cov=gendiff --cov-report xml:coverage.xml"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
